--- app/options/_lib/index.html
@@ -1,52 +1,0 @@
-<!DOCTYPE html>
-<html>
-	<head>
-		<meta charset="utf-8">
-		<title>環境設定 ｜ Narou Tweaker</title>
-		<meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no">
-		<base target="_blank">
-		<link rel="stylesheet" type="text/css" href="/common.css">
-		<link rel="stylesheet" type="text/css" href="../general.css">
-		<link rel="stylesheet" type="text/css" href="../novel/c_novel.css">
-		<link rel="stylesheet" type="text/css" href="/libs/font-awesome/css/all.css">
-		<link rel="stylesheet" type="text/css" href="/libs/Coloris/dist/coloris.min.css">
-		<link rel="stylesheet" type="text/css" href="/libs/highlightjs/styles/atom-one-light.min.css">
-		<script src="/libs/jquery-3.7.1.min.js"></script>
-		<script defer src="/libs/font-awesome/js/brands.min.js"></script>
-		<script defer src="/libs/font-awesome/js/fontawesome.min.js"></script>
-		<script src="/libs/SortableJS/Sortable.min.js"></script>
-		<script src="/libs/Coloris/dist/coloris.min.js"></script>
-		<script src="/libs/highlightjs/highlight.min.js"></script>
-		<script src="/libs/highlightjs/languages/css.min.js"></script>
-		<script src="/libs/highlightjs/languages/json.min.js"></script>
-		<script src="/libs/highlightjs/languages/xml.min.js"></script>
-		<script src="main.js" type="module"></script>
-        <input id="option-page-id" type="hidden" value="novel">
-    </head>
-    <body>
-        <div id="container">
-			<div id="sidebar"></div>
-			<div id="contents">
-				<div id="header">
-					<div id="header-title">
-						<div id="header-title--heading"></div>
-						<div id="header-title--description"></div>
-					</div>
-					<div id="header-menu">
-						<div class="header-menu-left">
-							<ul id="header-menu-left-items"></ul>
-						</div>
-					</div>
-				</div>
-
-                <div id="main"></div>
-
-                <div id="footer">
-                    <div id="js-failed">
-                        JavaScript failed to load. Your development build failed or your browser has some issue.
-                    </div>
-                </div>
-            </div>
-        </div>
-    </body>
-</html>
