--- app/options/kasasagi/c_kasasagi.js
@@ -1,10 +1,0 @@
-import { restoreOptions, setupDOM } from "../general.js";
-import { buttonHide, optionHide } from "../utils.js";
-
-setupDOM()
-buttonHide()
-optionHide()
-
-document.addEventListener('DOMContentLoaded', function(){
-    restoreOptions()
-})
