+++ app/options/mypage/main.js
@@ -1,0 +1,3 @@
+import { setup } from "../general.js";
+
+setup()
