+++ app/options/workspace/main.js
@@ -1,0 +1,6 @@
+import { setup } from "../general.js";
+import { setSortable, restoreHeaderIconList } from "../novel/header.js";
+
+setup()
+restoreHeaderIconList()
+setSortable()
